Testing tracking of ignored Sketch file
In this use case I've ignored all .sketch files and then created a new Sketch file from within Kactus. I changed it and am adding it to the repo to see if it transfers adequately without detailed tracking or if I have to manually regenerate the file on each machine.